+++ src/js/info-about-levels-game.js
@@ -1,0 +1,9 @@
+function levelsGame() {
+    window.levelsGame = {
+        light: 3,
+        medium: 6,
+        hard: 9,
+    };
+}
+
+export { levelsGame };
