+++ src/js/init-game-state.js
@@ -1,0 +1,36 @@
+import {
+    createForm,
+    createTimerForGame,
+    createButton,
+    createCards,
+} from './blocks';
+
+import { renderStartPage, renderGamePage } from './screens';
+
+function initGameState() {
+    return (window.application = {
+        blocks: {
+            ['create-form']: createForm,
+            ['create-timer-for-game']: createTimerForGame,
+            ['create-button']: createButton,
+            ['create-cards']: createCards,
+        },
+        screens: {
+            ['render-start-page']: renderStartPage,
+            ['render-game-page']: renderGamePage,
+        },
+        renderScreen: function (screenName) {
+            window.application.timers.forEach((time) => {
+                clearInterval(time);
+            });
+
+            this.screens[screenName]();
+        },
+        renderBlock: function (blockName, container) {
+            window.application.blocks[blockName](container);
+        },
+        timers: [],
+    });
+}
+
+export { initGameState };
